requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
PyYAML>=6.0.2
pytz>=2024.1
requests>=2.32.3
